# docs/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/upload/presigned-url": {
      "post": {
        "operationId": "UploadController_getPresignedUrl",
        "summary": "Generate a pre-signed URL for file upload to S3",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PresignedUrlRequestDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Pre-signed URL generated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PresignedUrlResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "tags": [
          "Upload"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
    "/teammates/requests": {
      "post": {
        "operationId": "TeammatesController_sendRequest",
        "summary": "Send team up request to another user",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SendTeammateRequestDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Request sent"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "User not found"
          },
          "409": {
            "description": "Already teammates or pending request exists"
          }
        },
        "tags": [
          "Teammates"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "get": {
        "operationId": "TeammatesController_getRequests",
        "summary": "Fetch teammate requests (incoming or outgoing)",
        "parameters": [
          {
            "name": "type",
            "required": false,
            "in": "query",
            "schema": {
              "default": "incoming",
              "enum": [
                "incoming",
                "outgoing"
              ],
              "type": "string"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "Page number",
            "schema": {
              "minimum": 1,
              "default": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Items per page",
            "schema": {
              "minimum": 1,
              "maximum": 50,
              "default": 20,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Paginated list of requests"
          }
        },
        "tags": [
          "Teammates"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/teammates/requests/{id}/respond": {
      "patch": {
        "operationId": "TeammatesController_respond",
        "summary": "Respond to a team up request (accept/decline/cancel)",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RespondTeammateRequestDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Request responded"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Only addressee can respond"
          },
          "404": {
            "description": "Request not found"
          },
          "409": {
            "description": "Request already responded"
          }
        },
        "tags": [
          "Teammates"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/teammates/count": {
      "get": {
        "operationId": "TeammatesController_getCount",
        "summary": "Get number of teammates",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Teammate count"
          }
        },
        "tags": [
          "Teammates"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/teammates": {
      "get": {
        "operationId": "TeammatesController_getTeammates",
        "summary": "Get teammates list (paginated)",
        "parameters": [
          {
            "name": "page",
            "required": true,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": true,
            "in": "query",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Paginated list of teammates"
          }
        },
        "tags": [
          "Teammates"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/teammates/suggestions": {
      "get": {
        "operationId": "TeammatesController_getSuggestions",
        "summary": "Get user suggestions for team up (same challenges, search by name)",
        "parameters": [
          {
            "name": "search",
            "required": false,
            "in": "query",
            "description": "Search by first name or last name",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "Page number",
            "schema": {
              "minimum": 1,
              "default": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Items per page",
            "schema": {
              "minimum": 1,
              "maximum": 50,
              "default": 20,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Paginated suggestions with common challenges"
          }
        },
        "tags": [
          "Teammates"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/notifications": {
      "get": {
        "operationId": "NotificationsController_getNotifications",
        "summary": "Get notifications (paginated)",
        "parameters": [
          {
            "name": "unread_only",
            "required": false,
            "in": "query",
            "description": "Filter to unread only",
            "schema": {
              "default": false,
              "type": "boolean"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "Page number",
            "schema": {
              "minimum": 1,
              "default": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Items per page",
            "schema": {
              "minimum": 1,
              "maximum": 50,
              "default": 20,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Paginated list of notifications"
          }
        },
        "tags": [
          "Notifications"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/notifications/unread-count": {
      "get": {
        "operationId": "NotificationsController_getUnreadCount",
        "summary": "Get unread notification count (for badge)",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Unread count"
          }
        },
        "tags": [
          "Notifications"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/notifications/read-all": {
      "patch": {
        "operationId": "NotificationsController_markAllAsRead",
        "summary": "Mark all notifications as read",
        "parameters": [],
        "responses": {
          "200": {
            "description": "All notifications marked as read"
          }
        },
        "tags": [
          "Notifications"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/notifications/{id}/read": {
      "patch": {
        "operationId": "NotificationsController_markAsRead",
        "summary": "Mark a notification as read",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Notification marked as read"
          },
          "404": {
            "description": "Notification not found"
          }
        },
        "tags": [
          "Notifications"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
  },
  "info": {
    "title": "Challenges.me API",
    "description": "Social habit and challenge platform API",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "PresignedUrlRequestDto": {
        "type": "object",
        "properties": {
          "fileName": {
            "type": "string",
            "description": "Original file name",
            "example": "photo.jpg"
          },
          "mimeType": {
            "type": "string",
            "description": "MIME type of the file",
            "example": "image/jpeg"
          },
          "folder": {
            "type": "string",
            "description": "Upload folder/category",
            "enum": [
              "profile",
              "checkin",
              "challenge",
              "message"
            ],
            "example": "checkin"
          }
        },
        "required": [
          "fileName",
          "mimeType",
          "folder"
        ]
      },
      "PresignedUrlResponseDto": {
        "type": "object",
        "properties": {
          "uploadUrl": {
            "type": "string",
            "description": "Pre-signed URL for uploading the file to S3"
          },
          "fileUrl": {
            "type": "string",
            "description": "Public URL via CloudFront to access the file after upload"
          },
          "key": {
            "type": "string",
            "description": "S3 storage key"
          }
        },
        "required": [
          "uploadUrl",
          "fileUrl",
          "key"
        ]
      },
...
      "SendTeammateRequestDto": {
        "type": "object",
        "properties": {
          "addressee_id": {
            "type": "string",
            "description": "User ID to send team up request to"
          }
        },
        "required": [
          "addressee_id"
        ]
      },
      "RespondTeammateRequestDto": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "pending",
              "accepted",
              "declined",
              "canceled",
              "blocked"
            ],
            "description": "Response: accepted, declined, or canceled"
          }
        },
        "required": [
          "status"
        ]
      },
...
    }
  }
}